swh.storage: Improve license endpoint's unknown license policy
Prior to this commit, the content_fossology_license_add endpoint was
not symmetric with other endpoints since it returned the list of
licenses it was unable to deal with. Now, it simply adds those unknown
licenses in the fossology_license table. And it no longer returns
anything (as other endpoints).

Related T721