archiver: Make package work for actual objstorage dependency version
The head version of the objstorage migrated. Thus the archiver tests
broke locally.

The test code was fixed to make the local tests pass with the actual
head version of the objstorage.

But we did not deploy the new objstorage yet, thus the archiver tests
break during the packaging phase (of any swh.storage.* package).

This commit is a temporary fix to make everyone happy (since we do
need to package other storage version for example).

This needs to be reverted as soon as we bump the dependency to the new
objstorage (> 0.17).

Related f681f83
Related b4dfe06
Related 08827dc
Related D205