Drop backwards-compatibility features between snapshots and occurrences
Summary:
Old occurrences have been fully migrated to snapshots now, so backwards
compatibility is not needed anymore. That's one more step towards the removal of
occurrences (T830).

This changes the origin_visit_get_by method to stop returning a list of
occurrences. The only user of this method is the api/1/origin/{x}/visit/{y} view
of the web API, which will need to be updated for T1207 anyway.

Test Plan:
integration tests updated for origin_visit_get_by; Running the web
app with the updated storage only yields the expected change in behavior for the
origin/visit API view, navigation is unaffected.

Reviewers: #reviewers!

Differential Revision: https://forge.softwareheritage.org/D466