Drop confusing functions revision_get_by/revision_log_by
Summary:
The combination of timestamp search for an occurrence with branch name filtering
never got ported to the snapshot model. As this function only has a single user,
which is a single endpoint deep in the public-facing API, let's just drop it.

revision_log_by uses the same filter but gives a log of parent revisions, and
has never been used by any client.

This will break this API endpoint, which is tracked in T1221, but iff we want to
keep it, I think it can be easily reimplemented by composing origin_visit_get_by
and the new snapshot_get_branches function.

Test Plan: It's dead, Jim.

Reviewers: zack

Differential Revision: https://forge.softwareheritage.org/D534