pgstorage: Empty temp tables instead of dropping them
Due to our pattern of adding objects [1], vacuum is triggered regularly on
pg_catalog.*, having an heavy impact on performance.

This commit tries to avoid the dropping the temporary tables part, emptying
them instead (they still are dropped at the end of the session but less often).
This should decrease the bloat on pg_catalog.* tables.

[1]

- create temporary table
- insert data from temporary table to production table with filtering
- drop temporary table