Adapt cassandra backend to validating model types
This is required to be able to activate type validation in the model (in
swh.model.model).

It requires to replace the "distorded" usage that was done of model entities
to build objects compatible with CqlRunner's object addition logic.
Since we cannot create invalid model entities any more in this context,
we add a new CassObject type (just a dict with __getattr__=__getitem__)
and use it as object passed to the CqlRunner for entity types that need
special care (namely Revision and Release).

This should still work with swh.model v0.0.62 (without type validation) as well
as the (next) v0.0.63 which will come with type valdation.