Refactor the JournalWriter class to normalize its methods
make all the add_xxx methods take an Iterable as argument. Also extract
the conditional execution of the underlying journal method calls in a
dedicated methof to prevent repeating the ``if self.journal`` all over the
place.

Ensure all the storage code that uses this JournalWriter class are updated
accordingly.