Adapt to swh.model 0.3
in which List attributes have replaced by Tuple ones.

This requires a bit of adaptation in the code of the
ValidatingProxyStorage to ensure dict representation of revision
objects are properly typed.

The test_api_client_dicts.py has been removed since it's not really
useful any more and would require a fair amount of work to fix it.