Deduplicate origin-metadata when they have the same authority + discovery_date + fetcher.
By replacing the old value with the new one.

This will allow an easy implementation of pagination, using the fetcher
id as an opaque page_token.

Plus, it did not make sense logically to have different metadata from
the same authority at the same time (especially with the same fetcher).