Add a new TenaciousProxyStorage
This proxy storage attempt to add buckets of objects, but in case of failure,
it splits the bucket in parts so every valid object in the bucket get a chance
to be inserted.

Also provides an error rate-limiting feature.

This proxy storage is mainly dedicated to help mirrorring an archive
using the replayer stack.