pytest_plugin: Avoid fixture client to declare optional dependency
Prior to this commit, this would make swh_storage_backend_config fixture clients
need to declare an optional dependency on swh.journal. Otherwise, it would not
work [1].

This commit fixes it by dropping this configuration in the main pytest plugin.
It keeps the storage tests testing with that journal_writer collaborator though
by declaring an override which still provides it.

This fixes the debian build [1]

[1] https://jenkins.softwareheritage.org/view/Debian%20packages/job/debian/job/packages/job/DLDBASE/job/gbp-buildpackage/152/console