test_storage: Make swh_contents fixture generate content model objects
Note that this will be improved upon after the storage migration. We can take a
look at the swh.model.tests.generate_testdata.gen_contents and make it generate
BaseContent objects directly if it's possible. This was not done here so the
impacts is limited to storage.

Related to T2494