Drop validate proxy
The validate proxy was initially an helper to ease the transition from the use
of dicts towards model objects in "*_add" production endpoints. It was not
removed immediately and grew some behavior it should not have (notably revision
conversion so the comparison within those related tests work).

After finally migrated away from dicts within the tests, we can now drop it [1].

Note that this moves the extra revision conversion behavior from the validate
proxy to those related tests. This extra step will also disappear when we
finally move the "*_get" endpoints to return model objects as well.

Note:
- This drops fixture redefinitions in the process (introduced so we could have
that validate proxy at the time).
- Remove the "validate" keyword from the get_storage function (so no longer
possible to instantiate one [2])

[1] T2994

[2] which, practically, is the case today, nothing runs on production with it.

Related to T2499