pg: Make date_neg_utc_offset is not null if date is not null.
Fixes a regression in 038a219f84d6b8a4f02b48f9ad3c5d823d097790,
as it made converters expect a boolean.

We stopped writing this kind of nulls since we started using model
objects for insertions, but didn't migrate existing data.

This commit migrates these nulls to false, then adds a constraint
to enforce it in the future.