pg: Check revision.extra_headers is not null.
Fixes a regression in 038a219f84d6b8a4f02b48f9ad3c5d823d097790,
as it made the converter expect a list.

When adding this column, we made it default to null instead of defaulting
to an empty array, so existing records were initialized will null.

This commit migrates these nulls to empty arrays, then adds a constraint
to enforce it in the future.