Add support for a new "check_config" config option in get_storage()
if "check_config" is present in get_storage()'s kwargs, call the
storage.check_config() method and raise an exception is the result
fails. The storage.check_config() method is called with the content
of the "check_config" config value (so expected value for this
currently is `check_config={"check_write": <bool>}`).