Finalize the config "local" deprecation in favor of "postgresql"
This will remove further deprecation warnings from the tests, especially the ones from
other modules depending on the storage's pytest-plugin.

This also fixes some edge case configuration for the backfill and the storage rpc
backend which would have been broken if we switched to that new name prior to this.

Related to b487a21f