Fix flakiness of test_origin_visit_status_get_random_nothing_found
start is increased from 13 to 14, because 13 weeks is 91 days,
ie. 30+31+30; so it is sometimes smaller than 3 months.

This was only hit rarely because the number of visits was small,
so this commit also increases the number of visits to make the
test more likely to fail if it should actually fail.