tests: only flush() the kafka journal writer once per test
instead of flushing it n times per test. Since the call to kafka
Producer.flush() takes about 1s, reducing the number of calls to this
method significantly reduce the execution time of the tests.

This required a small refactoring of the JournalBackfiller class to make
the journal writer live out of the scope of the run() so the test can
access the journal writer instance and call the flush() method.

Requires swh-journal >= 1.2.0