test_retry: Use proper way to mock sleep of retryable storage methods
Previous implementation was not mocking sleep of retryable storage methods
as the RetryingProxyStorage setup the retry features when it is instantiated.

So modify fixture to ensure sleep functions are mocked and return the mocks
in a dict indexed by storage method names.

This fixes debian buster package build for swh-storage.