Fix documentation build
It was broken by 0e8da810ac962b79dda18d9cfe9f3c7ea4f9de52
because the conditional import prevents sphinx from detecting
where the class is imported from.

Additionally, this attribute should not be on the interface,
because proxies do not have it, but they are expected to have
that interface.