setuptools: Include swh packages when building distribution archive
Currently, latest setuptools release makes the `find_packages` warns when building the
distribution archive. It mentions that some included folders sql, sql/upgrade, etc...
are imported but not explicitly declared as packages [1] [2]. As we only want the swh.*
module, explicitely mentions it.

According to the setuptools documentation [2], this now builds the archive as we want
it (as before including the swh.storage.{sql,tests,...} folders).

[1]
```
Python recognizes 'swh.storage.sql' as an importable package[^1],
10:32:14          but it is absent from setuptools' `packages` configuration.
...
```

[2] https://jenkins.softwareheritage.org/view/swh-debian%20(draft)/job/debian/job/packages/job/DSTO/job/gbp-buildpackage/468/console

[3] https://setuptools.pypa.io/en/latest/userguide/package_discovery.html#finding-namespace-packages

Refs. swh/infra/sysadm-environment#4997