Add StorageInterface.skipped_content_find()
To create recovery bundle, `swh-alter` needs to be able to retrieve
full SkippedContent objects from the storage.

The new method `skipped_content_find()` allows to retrieve all
SkippedContent objects matching a given set of hashes. Usually, this
should only be one, but multiple objects might be returned in case of
hash collisions.

While implementing this, swh-storage#4693 was identified which prevent
testing the implementation with the PostgreSQL storage in the case
a SkippedContent references a known origin.

Thanks to olasd and vlorentz for the reviews and suggesting small
improvements.