Add ObjectDeletionInterface aimed at Storage classes
swh-alter needs an interface in order to remove objects from the storage
(so we can handle takedown notices).

The chosen interface is optimal from swh-alter point of view: it
identifies a whole range of objects with different types that can be
removed alongside a given set of origins.

Giving all objects to be removed at once might also help with
consistency constraints inside the various storage facilities.

Only objects from this facility will be removed. The same method
should be called on other storage, objstorage, or journal instances
where the specified objects need to be removed.