replayer: add a --known-missmatched-hashes argument to the replay command
This accepts a file of swhids of objects that are known to be invalid
(hash mismatch) but should be replayed anyway (typically because they do
exist as is in the original storage).

The file is expected to have rows like:

  swh:1:xxx:<invalid_hex_hash>,<expected_hex_hash>
  [...]

Note that the cli only accepts swhids in the exception file, while the
backend (ModelObjectDeserializer) support all HashableObject. But we
currently do not need this feature on the cli tool for other object
types, and doing it this way is simpler in terms of type annotation.