postgresql: fix content_add_metadata()
it was using a generator like a list, effectively only inserting one
content object per batch.
The test needs to be modified to actually show the misbehavior, with the
need for a inserting first one object before inserting a list of
objects, otherwise the test is green without the fix (the reason is left
as an exercice for the curious reader).

Also fix a small type inconsistency on content_add() and use the same test
pattern in test_content_add().