Add command line tool to remove old object reference partition tables
`swh storage remove-old-object-reference-partitions 2023-09-01` can
be used to remove all partition tables for weeks before the given date.

By default, this will print the weeks for which tables would be dropped
and ask for a confirmation. The `--force` option will just proceed
directly.

The command will simply refuse to drop all partitions as it most
probably an error.