storage_tests: run test_types on the currently tested storage, not its backend
The intent behind test_types is to test the signature of wrapped
storages, to check that they match that of the StorageInterface
Protocol. However, the way the test was refactored ended up testing the
storage being *wrapped* by the storage under test, masking a few
inconsistencies in the way storages are being wrapped.

Unfortunately this breaks the tenacious proxy's test in an inscrutable
way (even when `functools.wraps`ing the return values of its
`__getattr__` function).