Introduce a PostgreSQL storage schema and API for object masking information
This is a simple database of the SWHIDs of objects for which we have
made a policy decision to restrict the diffusion without removing them
from the archive, and a lightweight history structure for the associated
object masking requests.

Doing this as an overlay, instead of modifying the storage schema for
all objects, allows us to start better separating the concerns of
archival of origins (which necessitates a full view of all the
unmodified objects that are stored in the archive), with the concerns
about the dissemination of said archived objects.

To avoid interfering with archival, the masking policy will only be
applied for full object retrieval and implemented as a new proxy
storage, which will be placed in front of all public-facing storages.