RevisionsWalker: Return Revision objects (and add support for revision_log returning them)
This is a breaking change, so we should release this as part of a new
major version (v3.0.0).

For now, this keeps support for existing (v2.x.x) backends returning dicts,
but adds support for v3.0.0 backends which will return objects.

This allows upgrading clients one by one, before upgrading servers.