Move SQL scripts for the masking proxy under swh/storage/proxies/masking
Make the MaskingProxy uses its own db rather than sharing the main
storage one with fragile use of flavors to shoehorn it in there.

The migration script for the main storage db requires to recreate the
db flavor table and (enum) type since this later cannot be altered
to remove an entry.

On the masking part, we set the db model version to 193 because the
currently deployed version is actually set to the main storage's db
version (thus 192).