Support removal of RawExtrinsicMetadata and ExtID objects
These changes are needed to allow `swh alter remove` to remove
RawExtrinsicMetadata and ExtID objects from the PostgreSQL and Cassandra
implementation of the storage.

As RawExtrinsicMetadata objects can be addressed using an
ExtendendSWHID, they can now be removed using `object_delete()`.

For ExtID, a new method `extid_delete_for_target()` has been added to
`ObjectDeletionInterface`. It will remove any ExtID object targeting one
of the object in the a list of SWHID given as an argument.

Address swh-alter!11