Introduce an object patching proxy feature in the MaskingProxy
This is an alternative to modifying the `person` table, as that
technique only worked on the PostgreSQL backend and not with Cassandra,
as it does not have a `person` table (authors and committers are inlined
in the revision and release tables).

This does not come with cli tooling to manage the display name table
entries for now.