Make the blocking/masking admin configs compatible with the swh db utilities
To be able to use the `swh db` command line, the configurations need to
have the following shape:

<toplevel key>:
  cls: postgresql
  db: <postgresql DSN>

The usage of `blocking_db` and `masking_db` instead of plain `db`, and
lack of overridability in the swh db utilities, made schema migrations
more annoying than necessary.

The old configuration key still works but raises a Deprecation warning.

The configuration of the proxies themselves is unchanged, as having a
differently-named key for the proxy-specific database makes sense to
avoid confusion.