Warn when objects are still buffered during shutdown
It is fairly easy to overlook calling `storage.flush()` after adding
some objects just before the end of a process. In a direct
configuration, everything will be fine but if a buffering proxy is
configured, this means that the objects still in the buffers are going
to be lost. Let’s add a warning if this is the case.