Support storing contents with arbitrary hash collisions
To avoid data loss on deployments with sha1-based objstorages, when
adding a Content object which is subject to a hash collision, the
content metadata is added to storage and the whole object (including raw
bytes) is sent to a new dedicated journal topic for further
processing (e.g. inclusion in a dedicated objstorage that supports
multi-hash addressing).

In practice, this commit removes the unique index constraints of the
PostgreSQL backend, and modifies the manual HashCollision exception
generation in the Cassandra/in-memory backends to handle the new journal
topic.