api/server: Consider QueryTimeout exception as a transient error
It has been observed in production that some QueryTimeout exceptions can be
raised by the cassandra server while querying the storage, especially when
cooking git bare repository with the vault.

Those errors are transient as usually the operations can be retried
successfully after a couple of attempts.

Ensure to return a TransientRemoteException instance for such exception to
inform the client it is a transient error and it can retry.

Related to swh/meta#5100.